+++ test/nextgen/operators/matmul/pack_rhs/matmul_pack_rhs_wrapper_registry.cpp
@@ -1,5 +1,6 @@
 //
 // SPDX-FileCopyrightText: Copyright 2025-2026 Arm Limited and/or its affiliates <open-source-office@arm.com>
+// SPDX-FileCopyrightText: Copyright 2026 Fujitsu Limited
 //
 // SPDX-License-Identifier: Apache-2.0
 //
@@ -10,14 +11,18 @@
 #include <array>
 #include <cstdint>
 #include <memory>
+#include <string>
+#include <string_view>
 #include <vector>
 
+#include "kai/ukernels/matmul/kai_matmul.h"
 #include "kai/ukernels/matmul/kai_matmul_pack_rhs.h"
 #include "kai/ukernels/matmul/matmul_clamp_f32_qai8dxp_qsi4cxp/kai_matmul_clamp_f32_qai8dxp1vlx4_qsi4cxp4vlx4_1vlx4vl_sme_mopa.h"
 #include "kai/ukernels/matmul/matmul_clamp_f32_qai8dxp_qsi4cxp/kai_matmul_clamp_f32_qai8dxp1vlx8_qsi4cxp4vlx8_1vlx4vl_sme2_mopa.h"
 #include "kai/ukernels/matmul/matmul_clamp_f32_qai8dxp_qsi4cxp/kai_matmul_clamp_f32_qai8dxp1x4_qsi4cxp4vlx4_1x4vl_sme2_sdot.h"
 #include "kai/ukernels/matmul/pack/kai_rhs_pack_kxn_f32p2vlx1biasf32_f32_f32_sme.h"
 #include "kai/ukernels/matmul/pack/kai_rhs_pack_nxk_qsi4cxps1s0_qsu4cxs1s0_neon.h"
+#include "kai/ukernels/matmul/pack/kai_rhs_pack_nxk_qsi8cxp_qsi8cx_neon.h"
 #include "test/common/data_type.hpp"
 #include "test/common/sme.hpp"
 #include "test/common/sve.hpp"
@@ -28,8 +33,10 @@
 #include "test/nextgen/format/plain_format.hpp"
 #include "test/nextgen/format/two_level_blockwise_format.hpp"
 #include "test/nextgen/harness/kernel_wrapper.hpp"
+#include "test/nextgen/operators/matmul/matmul_pack_args.hpp"
 #include "test/nextgen/operators/matmul/pack_rhs/matmul_pack_rhs_fp_nt_wrapper.hpp"
 #include "test/nextgen/operators/matmul/pack_rhs/matmul_pack_rhs_interface.hpp"
+#include "test/nextgen/operators/matmul/pack_rhs/matmul_pack_rhs_quant_i8_wrapper.hpp"
 #include "test/nextgen/operators/matmul/pack_rhs/matmul_pack_rhs_quant_wrapper.hpp"
 #include "test/nextgen/operators/matmul/pack_rhs/matmul_pack_rhs_ukerapi_t_wrapper.hpp"
 #include "test/nextgen/operators/matmul/pack_rhs/matmul_pack_rhs_ukerapi_wrapper.hpp"
@@ -37,6 +44,26 @@
 namespace kai::test {
 
 namespace {
+
+std::unique_ptr<KernelWrapper<MatShape>> create_matmul_rhs_pack_nxk_qsi8cxp_qsi8cx_neon(
+    std::string_view block_name, size_t block_height, size_t block_width) {
+    return std::make_unique<MatMulPackRhsQuantI8Wrapper>(
+        "matmul_rhs_pack_nxk_qsi8cxp" + std::string(block_name) + "_qsi8cx_neon",
+        MatMulPackRhsQuantI8Interface{
+            kai_get_n_step_rhs_pack_nxk_qsi8cxp_qsi8cx_neon,
+            kai_get_rhs_offset_rhs_pack_nxk_qsi8cxp_qsi8cx_neon,
+            kai_get_rhs_packed_stride_rhs_pack_nxk_qsi8cxp_qsi8cx_neon,
+            kai_get_rhs_packed_offset_rhs_pack_nxk_qsi8cxp_qsi8cx_neon,
+            kai_get_rhs_packed_size_rhs_pack_nxk_qsi8cxp_qsi8cx_neon,
+            kai_run_rhs_pack_nxk_qsi8cxp_qsi8cx_neon,
+        },
+        make_poly<PlainFormat>(DataType::I8), make_poly<PlainFormat>(DataType::FP32),
+        make_poly<PlainFormat>(DataType::FP32), make_poly<PlainFormat>(DataType::I32),
+        make_poly<Block2dRowFormat>(
+            block_height, block_width, 32, false, DataType::I8, std::array<DataType, 0>{},
+            std::array{DataType::I32, DataType::FP32, DataType::FP32}),
+        MatMulPackArgs{0, block_height, block_width, 1, 0});
+}
 
 bool portion_non_empty(
     size_t full_height, size_t full_width, size_t scheduler_block_height, size_t scheduler_block_width,
@@ -67,6 +94,18 @@ Poly<Format> unused_bias_format() {
 }
 
 }  // namespace
+
+std::unique_ptr<KernelWrapper<MatShape>> create_matmul_rhs_pack_nxk_qsi8cxp8x4_qsi8cx_neon() {
+    return create_matmul_rhs_pack_nxk_qsi8cxp_qsi8cx_neon("8x4", 8, 4);
+}
+
+std::unique_ptr<KernelWrapper<MatShape>> create_matmul_rhs_pack_nxk_qsi8cxp32x4_qsi8cx_neon() {
+    return create_matmul_rhs_pack_nxk_qsi8cxp_qsi8cx_neon("32x4", 32, 4);
+}
+
+std::unique_ptr<KernelWrapper<MatShape>> create_matmul_rhs_pack_nxk_qsi8cxp8x8_qsi8cx_neon() {
+    return create_matmul_rhs_pack_nxk_qsi8cxp_qsi8cx_neon("8x8", 8, 8);
+}
 
 std::unique_ptr<KernelWrapper<MatShape>> create_matmul_rhs_pack_nxk_qsi4cxp4vlx4s1s0_qsu4cxs1s0_neon() {
     return std::make_unique<MatMulPackRhsQuantWrapper>(
@@ -338,6 +377,48 @@ std::unique_ptr<KernelWrapper<MatShape>> create_matmul_rhs_pack_nxk_qsi4c32p16vs
             MatMulSlot::RHS_T_QSCALE_RESCALED, MatMulSlot::RHS_T_QDATA_SIGN_SUM_SCALED, MatMulSlot::ACC_BIAS_N_DATA,
             MatMulSlot::RHS_T_QDATA_SIGN},
         MatMulSlot::RHS_T_QDATA, format_config);
+}
+
+bool is_shape_suitable_rhs_qai8dxp1x4_qsi8cxp8x4_1x8_sve_dot(
+    [[maybe_unused]] size_t shape_m, size_t shape_n, size_t shape_k, const MatrixPortion& portion) {
+    if (shape_n == 0 || shape_k == 0) {
+        return false;
+    }
+
+    const auto api = kai_matmul_clamp_f32_qai8dxp1x4_qsi8cxp8x4_1x8_sve_dot();
+    const kai_matmul_uker_config config{};
+    const size_t nr = api.get_step(&config).n;
+    const size_t rhs_n_step = kai_get_n_step_rhs_pack_nxk_qsi8cxp_qsi8cx_neon(nr);
+
+    return portion_non_empty(shape_n, shape_k, rhs_n_step, shape_k, portion);
+}
+
+bool is_shape_suitable_rhs_qai8dxp1x8_qsi8cxp8x8_1x8_sve_dot(
+    [[maybe_unused]] size_t shape_m, size_t shape_n, size_t shape_k, const MatrixPortion& portion) {
+    if (shape_n == 0 || shape_k == 0) {
+        return false;
+    }
+
+    const auto api = kai_matmul_clamp_f32_qai8dxp1x8_qsi8cxp8x8_1x8_sve_dot();
+    const kai_matmul_uker_config config{};
+    const size_t nr = api.get_step(&config).n;
+    const size_t rhs_n_step = kai_get_n_step_rhs_pack_nxk_qsi8cxp_qsi8cx_neon(nr);
+
+    return portion_non_empty(shape_n, shape_k, rhs_n_step, shape_k, portion);
+}
+
+bool is_shape_suitable_rhs_qai8dxp1x4_qsi8cxp32x4_1x32_sve_dot(
+    [[maybe_unused]] size_t shape_m, size_t shape_n, size_t shape_k, const MatrixPortion& portion) {
+    if (shape_n == 0 || shape_k == 0) {
+        return false;
+    }
+
+    const auto api = kai_matmul_clamp_f32_qai8dxp1x4_qsi8cxp32x4_1x32_sve_dot();
+    const kai_matmul_uker_config config{};
+    const size_t nr = api.get_step(&config).n;
+    const size_t rhs_n_step = kai_get_n_step_rhs_pack_nxk_qsi8cxp_qsi8cx_neon(nr);
+
+    return portion_non_empty(shape_n, shape_k, rhs_n_step, shape_k, portion);
 }
 
 bool is_shape_suitable_rhs_qai8dxp1vlx8_qsi4cxp4vlx8_1vlx4vl_sme2_mopa(